+++ src/main/java/com/pzhu/novel/service/ProxySpiderXpathConfigService.java
@@ -1,8 +1,8 @@
 package com.pzhu.novel.service;
 
-import java.util.List;
-
 import com.pzhu.novel.mbg.model.ProxySpiderXpathConfig;
+
+import java.util.List;
 
 /**
  * @author 刘鹏 liupeng
@@ -13,7 +13,7 @@ public interface ProxySpiderXpathConfigService {
 
     void update(ProxySpiderXpathConfig proxySpiderXpathConfig);
 
-    void delete(Integer id);
+    void delete(Long id);
 
     List<ProxySpiderXpathConfig> pageList(Integer pageNum, Integer pageSize, ProxySpiderXpathConfig proxySpiderXpathConfig);
 }
